Fix circular dependency by extracting interaction constants to separate file
- Create interactionConstants.ts to hold all dashboard interaction constants
- Update SceneTimeRange, variables, and components to import from new location
- Re-export constants from index files for backward compatibility
- Remove unused imports from SceneRenderProfiler
- Resolves 'Cannot access before initialization' errors in tests
