+++ packages/scenes/src/behaviors/SceneRenderProfiler.ts
@@ -1,10 +1,9 @@
 import { writePerformanceLog } from '../utils/writePerformanceLog';
-import { SceneQueryControllerLike } from './types';
+import { SceneQueryControllerLike, LongFrameEvent, SceneComponentInteractionEvent } from './types';
 import { getScenePerformanceTracker, generateOperationId, PerformanceEventData } from './ScenePerformanceTracker';
 import { PanelProfilingManager, PanelProfilingConfig } from './PanelProfilingManager';
 import { SceneObject } from '../core/types';
 import { writeSceneLog } from '../utils/writeSceneLog';
-import { SceneQueryControllerLike, LongFrameEvent, SceneComponentInteractionEvent } from './types';
 import { LongFrameDetector } from './LongFrameDetector';
 
 const POST_STORM_WINDOW = 2000; // Time after last query to observe slow frames
@@ -501,14 +500,4 @@ export function calculateNetworkTime(requests: PerformanceResourceTiming[]): num
   return totalNetworkTime;
 }
 
-export const REFRESH_INTERACTION = 'refresh';
-export const TIME_RANGE_CHANGE_INTERACTION = 'time_range_change';
-export const FILTER_ADDED_INTERACTION = 'filter_added';
-export const FILTER_REMOVED_INTERACTION = 'filter_removed';
-export const FILTER_CHANGED_INTERACTION = 'filter_changed';
-export const FILTER_RESTORED_INTERACTION = 'filter_restored';
-export const VARIABLE_VALUE_CHANGED_INTERACTION = 'variable_value_changed';
-export const SCOPES_CHANGED_INTERACTION = 'scopes_changed';
-export const ADHOC_KEYS_DROPDOWN_INTERACTION = 'adhoc_keys_dropdown';
-export const ADHOC_VALUES_DROPDOWN_INTERACTION = 'adhoc_values_dropdown';
-export const GROUPBY_DIMENSIONS_INTERACTION = 'groupby_dimensions';
+// Constants moved to ./interactionConstants.ts to avoid circular dependencies
