+++ packages/scenes/src/behaviors/index.ts
@@ -15,3 +15,18 @@ export {
   type PanelPerformanceData,
   type QueryPerformanceData,
 } from './ScenePerformanceTracker';
+
+// Interaction constants for performance tracking
+export {
+  REFRESH_INTERACTION,
+  TIME_RANGE_CHANGE_INTERACTION,
+  FILTER_ADDED_INTERACTION,
+  FILTER_REMOVED_INTERACTION,
+  FILTER_CHANGED_INTERACTION,
+  FILTER_RESTORED_INTERACTION,
+  VARIABLE_VALUE_CHANGED_INTERACTION,
+  SCOPES_CHANGED_INTERACTION,
+  ADHOC_KEYS_DROPDOWN_INTERACTION,
+  ADHOC_VALUES_DROPDOWN_INTERACTION,
+  GROUPBY_DIMENSIONS_INTERACTION,
+} from './interactionConstants';
