+++ packages/scenes/src/components/SceneRefreshPicker.tsx
@@ -8,7 +8,7 @@ import { SceneObjectBase } from '../core/SceneObjectBase';
 import { sceneGraph } from '../core/sceneGraph';
 import { SceneComponentProps, SceneObject, SceneObjectState, SceneObjectUrlValues } from '../core/types';
 import { SceneObjectUrlSyncConfig } from '../services/SceneObjectUrlSyncConfig';
-import { REFRESH_INTERACTION } from '../behaviors/SceneRenderProfiler';
+import { REFRESH_INTERACTION } from '../behaviors/interactionConstants';
 
 export const DEFAULT_INTERVALS = ['5s', '10s', '30s', '1m', '5m', '15m', '30m', '1h', '2h', '1d'];
