+++ packages/scenes/src/variables/adhoc/AdHocFiltersCombobox/AdHocFiltersCombobox.tsx
@@ -52,7 +52,7 @@ import {
   FILTER_CHANGED_INTERACTION,
   ADHOC_KEYS_DROPDOWN_INTERACTION,
   ADHOC_VALUES_DROPDOWN_INTERACTION,
-} from '../../../behaviors/SceneRenderProfiler';
+} from '../../../behaviors/interactionConstants';
 import { getInteractionTracker } from '../../../core/sceneGraph/getInteractionTracker';
 
 interface AdHocComboboxProps {
